+++ src/components/ExpertItem/index.module.less
@@ -18,6 +18,11 @@
     font-size: 26px;
     margin: 6px 0;
     color: #666;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    width: 100%;
+    text-align: center;
   }
   &__num {
     font-size: 20px;
