+++ src/pages/course/index.module.less
@@ -8,3 +8,33 @@
   font-size: 30px;
   background: #eceded;
 }
+.my__attention {
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  font-size: 30px;
+  width: 100%;
+  background: #eceded;
+  padding-top: 20px;
+  &__title {
+    background: #fff;
+    margin: 0 20px;
+    display: flex;
+    padding: 30px;
+    border-radius: 15px;
+    font-weight: bold;
+  }
+  &__li {
+    background: #fff;
+    margin: 20px;
+  }
+  &__expert__li {
+    background: #fff;
+    margin: 20px;
+    display: flex;
+    flex-wrap: wrap;
+  }
+  &__expert__li_empty {
+    display: block;
+  }
+}
