+++ src/pages/mine/components/AboutList/index.module.less
@@ -142,7 +142,7 @@
     &__txt {
       font-size: 26px;
       color: #666;
-      margin-left: 20px;
+      border: 0;
       font-weight: lighter;
     }
   }
