+++ src/pages/recommend/expertDetail/index.module.less
@@ -18,6 +18,7 @@
     .overview__img {
       width: 110px;
       height: 110px;
+      border: 1px solid #f8c4c4;
     }
     .overview__info {
       flex: 1;
